+++ src/PSDocs/Models/Include.cs
@@ -1,0 +1,9 @@
+﻿namespace PSDocs.Models
+{
+    public sealed class Include : DocumentNode
+    {
+        public override DocumentNodeType Type => DocumentNodeType.Include;
+
+        public string Path { get; set; }
+    }
+}